backend/requirements.txt
Flask==2.2.2
python-dotenv==0.19.1
mysql-connector-python==8.0.26
